style: updated styling for login,signup pages
