+++ src/pages/index.js
@@ -3,8 +3,8 @@ import {Products} from './Products/products';
 import {Wishlist} from './Wishlist/wishlist';
 import {Cart} from './cart/cart';
 import { ProductDetail } from './productDetail';
-import {Login} from './login';
-import { SignUp } from './signUp';
+import {Login} from './login/login';
+import { SignUp } from './login/signUp';
 import {Error} from './error';
 import {AccountDetails} from './AccountDetails/accountDetails';
 import {AddressDetails} from './AccountDetails/addressDetails';
