+++ src/pages/login/login.css
@@ -1,0 +1,35 @@
+.login-page{
+  margin: 4rem auto;
+}
+
+.form-container{
+  margin: auto;
+ box-shadow: 0 7px 29px 0 rgba(100,100,111,.2);
+ width: 30rem;
+ padding: 2rem;
+ font-size: 18px;
+ border-radius: 1rem;
+ background-color: #f8f8f8;
+}
+
+.form-container > * {
+  display: block;
+  margin: 30px auto;
+}
+
+.form-input {
+  width: 10rem;
+  text-align: left;
+}
+
+.form-input input{
+  font-size: 16px;
+  width: 15rem;
+  border-radius: 2px;
+  border: 2px solid #384c54;
+  padding: 8px
+}
+
+.new-account{
+  text-decoration: underline;
+}
